requirements.txt
yfinance
pandas
numpy
matplotlib
scipy
statsmodels
scikit-learn
pytest
seaborn
pandas_ta
dash
plotly
tensorflow
alpaca_trade_api
python-dotenv
